Use double precision for text size in export interface
Changed ExportInterface.exportAdvText sizex/sizey from int to double,
eliminating the integer rounding that caused up to 16% font size
mismatch for small text. Analysis of 216k text elements across 5000
drawings shows: >1% mismatch dropped from 68% to 18%, >5px error
from 14% to 1.7%, and 82% of elements now match within 1%.

+++ src/fidocadj/export/ExportEPS.java
@@ -217,7 +217,7 @@ public final class ExportEPS implements ExportInterface, TextInterface
         @throws IOException when things goes horribly wrong, for example if
             the file in which the output is being done is not accessible.
     */
-    public void exportAdvText (int x, int y, int sizex, int sizey,
+    public void exportAdvText (int x, int y, double sizex, double sizey,
         String fontnameT, boolean isBold, boolean isMirrored, boolean isItalic,
         int orientation, int layer, String textT)
         throws IOException
